Refactor AttestationPayload and related logic to use uint8 for signaturesRequired
- Changed the type of signaturesRequired in AttestationPayload from uint32 to uint8 for better memory efficiency.
- Updated related scripts and tests to accommodate the new type, ensuring compatibility across the codebase.
- Enhanced README.md to clarify the acceptance of Solana node manifest for cross-VM tests.
- Adjusted message construction logic in VerifierLib and MolphaSigLib to reflect the updated data structure.
- Updated test fixtures and assertions to align with the new signaturesRequired type.

+++ script/GenerateAttestationFixture.s.sol
@@ -21,7 +21,7 @@ contract GenerateAttestationFixture is Script {
 
     uint256 internal constant NODE_COUNT = 12;
     uint256 internal constant REDUNDANCY_BUFFER = 2;
-    uint32 internal constant SIGNATURES_REQUIRED = 5;
+    uint8 internal constant SIGNATURES_REQUIRED = 5;
     uint64 internal constant CANONICAL_TIMESTAMP = 1_700_000_000;
     bytes32 internal constant SOURCE_ID = keccak256("MOLPHA_CONSUMER_SDK_SOURCE");
